+++ maestro-cli/src/main/java/maestro/cli/session/MaestroSessionManager.kt
@@ -73,6 +73,9 @@ object MaestroSessionManager {
         isStudio: Boolean = false,
         isHeadless: Boolean = false,
         screenSize: String? = null,
+        extensions: List<String>? = null,
+        chromeProfile: String? = null,
+        profileDirectory: String? = null,
         reinstallDriver: Boolean = true,
         deviceIndex: Int? = null,
         executionPlan: WorkspaceExecutionPlanner.ExecutionPlan? = null,
@@ -116,6 +119,9 @@ object MaestroSessionManager {
             isStudio = isStudio,
             isHeadless = isHeadless,
             screenSize = screenSize,
+            extensions = extensions,
+            chromeProfile = chromeProfile,
+            profileDirectory = profileDirectory,
             driverHostPort = driverHostPort,
             reinstallDriver = reinstallDriver,
             platformConfiguration = executionPlan?.workspaceConfig?.platform
@@ -197,6 +203,9 @@ object MaestroSessionManager {
         isStudio: Boolean,
         isHeadless: Boolean,
         screenSize: String?,
+        extensions: List<String>? = null,
+        chromeProfile: String? = null,
+        profileDirectory: String? = null,
         reinstallDriver: Boolean,
         driverHostPort: Int?,
         platformConfiguration: PlatformConfiguration? = null,
@@ -220,7 +229,7 @@ object MaestroSessionManager {
                         platformConfiguration = platformConfiguration
                     )
 
-                    Platform.WEB -> pickWebDevice(isStudio, isHeadless, screenSize)
+                    Platform.WEB -> pickWebDevice(isStudio, isHeadless, screenSize, extensions, chromeProfile, profileDirectory)
                 },
                 device = selectedDevice.device,
             )
@@ -249,7 +258,7 @@ object MaestroSessionManager {
             )
 
             selectedDevice.platform == Platform.WEB -> MaestroSession(
-                maestro = pickWebDevice(isStudio, isHeadless, screenSize),
+                maestro = pickWebDevice(isStudio, isHeadless, screenSize, extensions, chromeProfile, profileDirectory),
                 device = null
             )
 
@@ -445,8 +454,15 @@ object MaestroSessionManager {
         )
     }
 
-    private fun pickWebDevice(isStudio: Boolean, isHeadless: Boolean, screenSize: String?): Maestro {
-        return Maestro.web(isStudio, isHeadless, screenSize)
+    private fun pickWebDevice(
+        isStudio: Boolean,
+        isHeadless: Boolean,
+        screenSize: String?,
+        extensions: List<String>? = null,
+        chromeProfile: String? = null,
+        profileDirectory: String? = null,
+    ): Maestro {
+        return Maestro.web(isStudio, isHeadless, screenSize, extensions, chromeProfile, profileDirectory)
     }
 
     private data class SelectedDevice(