+++ maestro-client/src/main/java/maestro/Maestro.kt
@@ -679,6 +679,9 @@ class Maestro(
             isStudio: Boolean,
             isHeadless: Boolean,
             screenSize: String?,
+            extensions: List<String>? = null,
+            chromeProfile: String? = null,
+            profileDirectory: String? = null,
         ): Maestro {
             // Check that JRE is at least 11
             val version = System.getProperty("java.version")
@@ -695,6 +698,9 @@ class Maestro(
                 isStudio = isStudio,
                 isHeadless = isHeadless,
                 screenSize = screenSize,
+                extensions = extensions,
+                chromeProfile = chromeProfile,
+                profileDirectory = profileDirectory,
             )
             driver.open()
             return Maestro(driver)